Expose Activity Stream endpoints to the public internet
To allow the new data-flow query the Activity Stream endpoints from outside of
GOV.UK PaaS, this change removes the check on the X-Forwarded-For header that
forbids any requests that contain such a header, which are the ones from the
public internet.

The 3 endpoints are at the paths:

/api/v1/activity-stream/, powered by ActivityStreamViewSet
/api/v1/activity-stream/users/, powered by ActivityStreamDirectorySSOUsers
/api/v1/activity-stream/user-answers-vfm/, powered by ActivityStreamDirectorySSOUserAnswersVFM

To mitigate the increased risk by the exposure, an IP filter should be added in
front of directory-sso (if it's not already running) should only allow the IP
address of the new data-flow into these endpoints.